% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\name{blacklist}
\alias{blacklist}
\title{Returns a list of disallowed functions.}
\usage{
blacklist()
}
\value{
list
}
\description{
\code{blacklist} returns a list of functions that are not allowed
to be executed. The package name containing the function is a
key the value of this key is a list of the names of all functions
in this package that are disallowed.
}
