% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildJobsList.R
\name{buildJobsList}
\alias{buildJobsList}
\title{This is a generic function for bundling a job in the format required by the workers.}
\usage{
buildJobsList(func, ..., resultsQueue, errorQueue)
}
\arguments{
\item{func}{The function to be executed.}

\item{...}{Parameters to be passed to \code{func}.}

\item{resultsQueue}{The queue to store returned results in.}

\item{errorQueue}{The queue to store any returned errors in.}
}
\description{
\code{buildJobsList} takes a function, a result key, and any parameters the
function requires and returns a list in the form the workers can understand, i.e.
a list with the following keys: Function, Parameters, and ResultsKey.
}
\details{
If the job you wish to run requires significant setup to define a job, it may be useful
to write a custom function to handle the construction of a job. This job can be passed
in place of \code{buildJobsList} in the \code{queueJobs} function.
}
