% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCallback.R
\name{defineCallback}
\alias{defineCallback}
\title{An internal function that takes a channel definition and creates the callback.}
\usage{
defineCallback(channelDef, envir)
}
\arguments{
\item{channelDef}{A list containing \code{channel} and \code{callback}, returned
from a channel definition function.}

\item{envir}{An environment object.}
}
\value{
The name of the channel to listen on.
}
\description{
The \code{redisSubscribe} function expects callbacks to a function with the same
name as the subscribed channel. \code{defineCallback} takes the callback
returned from a channel definition and assigns it to the correct function name
so \code{redisSubscribe} can access it.
}

