% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineChannels.R
\name{defineChannels}
\alias{defineChannels}
\title{A function that defines the channel callbacks and returns the channels to listen on.}
\usage{
defineChannels(channels, envir)
}
\arguments{
\item{channels}{A list of functions defining channels to subscribe to with
corresponding callbacks.}

\item{envir}{The environment to assign the callbacks in.}
}
\description{
\code{defineChannels} takes a list of functions defining channels. It creates the
appropriate callbacks on the global environment and then returns the names of the
channels to be passed later to \code{redisSubscribe}.
}
\details{
This function should be used if you plan on manually subscribing to and monitoring
channels. If you wish to automatically subscribe to the channels after they are
defined use \code{minionListener}.
}

