% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlplyQueueJobs.R
\name{dlplyQueueJobs}
\alias{dlplyQueueJobs}
\title{This function constructs all jobs and adds them to the specified queue.}
\usage{
dlplyQueueJobs(host, iter, variables, func, ..., resultsQueue, errorQueue,
  buildJobsList = buildJobsList, port = 6379, jobsQueue = "jobsqueue")
}
\arguments{
\item{host}{The name or ip address of the redis server.}

\item{iter}{Data frame to be iterated over.}

\item{variables}{Per \code{plyr} docs, variables to split data frame by, as
as.quoted variables, a formula or character vector.}

\item{func}{The function to be executed}

\item{...}{Parameters to be passed to \code{func}.}

\item{resultsQueue}{The queue to store returned results in.}

\item{errorQueue}{The queue to store any returned errors in.}

\item{buildJobsList}{The function that performs any required tasks and returns
job list with keys Function, Params, and ResultsKey for workers. Defaults
to generic \code{buildJobsList} function.}

\item{port}{The port the redis server is running on. Defaults to 6379.}

\item{jobsQueue}{A string giving the name of the queue where jobs will be placed.
Defaults to \code{jobsqueue}.}
}
\description{
For each iterator, \code{dlplyQueueJobs} takes the function to be executed
and any parameters the function needs, bundles them into a job list, and
pushes them to the specified queue.
}
\details{
\code{iter} is a data frame to iterate over. Use \code{variables} to
specify how to split it.
}

