% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessage.R
\name{getMessage}
\alias{getMessage}
\title{A function that takes a queue and a serialization type and returns the parsed message.}
\usage{
getMessage(conn, queue, useJSON = F, blocking = F)
}
\arguments{
\item{conn}{An open redux hiredis connection.}

\item{queue}{A string giving the name of the queue to get the message from.}

\item{useJSON}{Flag specifying whether messages are in JSON format. Defaults to false.}

\item{blocking}{Flag specifying whether a blocking pop or regular pop will be performed
when getting a message from the queue.}
}
\description{
\code{getMessage} takes a Redis connection, a string queue, and a boolean specifying
whether the serialization was JSON or not. It then gets the message from the queue,
parses it, and returns the parsed message.
}
