% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minionListener.R
\name{minionListener}
\alias{minionListener}
\title{A function that subscribes to a channel and performs various server-wide R tasks.}
\usage{
minionListener(host, port, channels, logging = T,
  logFileDir = "/var/log/R/")
}
\arguments{
\item{host}{The name or ip address of the redis server.}

\item{port}{The port the redis server is running on.}

\item{channels}{A list of functions defining channels to subscribe to with
corresponding callbacks.}

\item{logging}{A boolean to enable or disable logging to a file on the system. Defaults
to \code{true}.}

\item{logFileDir}{A string giving the directory to store worker log files if logging is
enabled.}
}
\description{
\code{minionListener} is a blocking function that will completely block the R
instance running it. It immediately performs some housekeeping by gathering
logging information and subscribing to the redis channel. Once subscribed, it
defines a callback function for the subscribed channel then monitors the channel
for messages to pass into the callback function and execute. This function is
primarily for installing packages on all minion servers simultaneously, but can
also be adapted to other server maintenance tasks.
}
\details{
Since \code{minionListener} uses the Redis PUB/SUB system instead of the POP/PUSH
system that \code{minionWorker} uses. \code{minionListener} will receive all
messages published on the channel(s) it is subscribed to and will execute the
callback on them. Due to the blocking nature of R, it is possible that published
messages could be missed if the listener is in the middle of executing the
callback on a previous message. Thus it is recommended that you allow sufficient
processing time between publishing messages.

Since each listener receives all messages on the subscribed channel, only one
listener needs to be run on each server.
}

